requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
backoff==1.11.1
black==22.3.0
broadcaster==0.2.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
fastapi==0.75.1
fastapi-websocket-rpc==0.1.21
frozenlist==1.3.0
future==0.18.2
gcloud-aio-auth==4.0.0
gcloud-aio-pubsub==5.0.0
gcloud-rest-auth==4.0.0
gcloud-rest-pubsub==5.0.0
google-api-core==2.7.1
google-auth==2.6.3
google-cloud-pubsub==2.12.0
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.4
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.13.0
httptools==0.4.0
idna==3.3
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tenacity==8.0.1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.2
yarl==1.7.2
